Whoopsie. Int to Float didn't work. Ended up with 100% or 0%.

+++ main.cpp
@@ -95,7 +95,7 @@ int main()
             maxTemp = runTemp;
 
         Difference = maxTemp - minTemp;
-        Percent    = (runTemp - minTemp)/Difference;
+        Percent    = (float)(runTemp - minTemp)/Difference;
         onTime     = Scalar * Percent;
         
         setPWM();
